do not create symlinks if user skips

+++ Rakefile
@@ -24,9 +24,10 @@ task :install do
         when 'O' then overwrite_all = true
         when 'B' then backup_all = true
         when 'S' then skip_all = true
-        when 's' then next
+        when 's' then skip = true
         end
       end
+      next if skip || skip_all
       FileUtils.rm_rf(target) if overwrite || overwrite_all
       `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
     end
